cnat: support per snat return session next node
Type: improvement

+++ src/plugins/cnat/cnat_node_feature.c
@@ -88,7 +88,7 @@ cnat_input_feature_new_flow_inline (vlib_main_t *vm, vlib_buffer_t *b,
   ts->ts_rw_bm |= 1 << CNAT_LOCATION_INPUT;
 
   rw->cts_lbi = (u32) ~0;
-  rw->cts_dpoi_next_node = (u32) ~0;
+  rw->cts_dpoi_next_node = (u16) ~0;
 
   cnat_make_buffer_5tuple (b, af, &rw->tuple, 0 /* iph_offset */,
 			   0 /* swap */);
@@ -141,7 +141,7 @@ cnat_input_feature_new_flow_inline (vlib_main_t *vm, vlib_buffer_t *b,
       ts->ts_rw_bm |= 1 << (CNAT_IS_RETURN + CNAT_LOCATION_OUTPUT);
 
       rrw->cts_lbi = (u32) ~0;
-      rrw->cts_dpoi_next_node = (u32) ~0;
+      rrw->cts_dpoi_next_node = (u16) ~0;
 
       cnat_make_buffer_5tuple (b, af, &rrw->tuple, 0 /* iph_offset */,
 			       1 /* swap */);
@@ -384,7 +384,7 @@ cnat_output_feature_new_flow_inline (vlib_main_t *vm, vlib_buffer_t *b,
   ts->ts_rw_bm |= 1 << (CNAT_LOCATION_OUTPUT);
 
   rw->cts_lbi = (u32) ~0;
-  rw->cts_dpoi_next_node = (u32) ~0;
+  rw->cts_dpoi_next_node = (u16) ~0;
   rw->fib_index = (u32) ~0;
 
   cnat_make_buffer_5tuple (b, af, &rw->tuple, iph_offset, 0 /* swap */);
@@ -439,7 +439,7 @@ cnat_output_feature_new_flow_inline (vlib_main_t *vm, vlib_buffer_t *b,
   ts->ts_rw_bm |= 1 << (CNAT_IS_RETURN + CNAT_LOCATION_INPUT);
 
   rrw->cts_lbi = (u32) ~0;
-  rrw->cts_dpoi_next_node = (u32) ~0;
+  rrw->cts_dpoi_next_node = (u16) ~0;
   rrw->fib_index = AF_IP4 == af ? cpe->ret_fib_index4 : cpe->ret_fib_index6;
 
   cnat_make_buffer_5tuple (b, af, &rrw->tuple, iph_offset, 1 /* swap */);
